+++ cvat-canvas/src/typescript/interactionHandler.ts
@@ -17,38 +17,25 @@ export interface InteractionHandler {
     transform(geometry: Geometry): void;
     interact(interactData: InteractionData): void;
     configurate(config: Configuration): void;
+    destroy(): void;
     cancel(): void;
 }
 
 export class InteractionHandlerImpl implements InteractionHandler {
     private onInteraction: (shapes: InteractionResult[] | null, shapesUpdated?: boolean, isDone?: boolean) => void;
-
     private configuration: Configuration;
-
     private geometry: Geometry;
-
     private canvas: SVG.Container;
-
     private interactionData: InteractionData;
-
     private cursorPosition: { x: number; y: number };
-
     private shapesWereUpdated: boolean;
-
     private interactionShapes: SVG.Shape[];
-
     private currentInteractionShape: SVG.Shape | null;
-
     private crosshair: Crosshair;
-
     private threshold: SVG.Rect | null;
-
     private thresholdRectSize: number;
-
     private intermediateShape: PropType<InteractionData, 'intermediateShape'>;
-
     private drawnIntermediateShape: SVG.Shape;
-
     private thresholdWasModified: boolean;
 
     private prepareResult(): InteractionResult[] {
@@ -375,6 +362,27 @@ export class InteractionHandlerImpl implements InteractionHandler {
         return false;
     }
 
+    private onKeyUp = (e: KeyboardEvent): void => {
+        if (this.interactionData.enabled && e.keyCode === 17) {
+            if (this.interactionData.onChangeToolsBlockerState && !this.thresholdWasModified) {
+                this.interactionData.onChangeToolsBlockerState('keyup');
+            }
+            if (this.shouldRaiseEvent(false)) {
+                // 17 is ctrl
+                this.onInteraction(this.prepareResult(), true, false);
+            }
+        }
+    };
+
+    private onKeyDown = (e: KeyboardEvent): void => {
+        if (!e.repeat && this.interactionData.enabled && e.keyCode === 17) {
+            if (this.interactionData.onChangeToolsBlockerState && !this.thresholdWasModified) {
+                this.interactionData.onChangeToolsBlockerState('keydown');
+            }
+            this.thresholdWasModified = false;
+        }
+    };
+
     public constructor(
         onInteraction: (
             shapes: InteractionResult[] | null,
@@ -452,26 +460,8 @@ export class InteractionHandlerImpl implements InteractionHandler {
             }
         });
 
-        window.addEventListener('keyup', (e: KeyboardEvent): void => {
-            if (this.interactionData.enabled && e.keyCode === 17) {
-                if (this.interactionData.onChangeToolsBlockerState && !this.thresholdWasModified) {
-                    this.interactionData.onChangeToolsBlockerState('keyup');
-                }
-                if (this.shouldRaiseEvent(false)) {
-                    // 17 is ctrl
-                    this.onInteraction(this.prepareResult(), true, false);
-                }
-            }
-        });
-
-        window.addEventListener('keydown', (e: KeyboardEvent): void => {
-            if (!e.repeat && this.interactionData.enabled && e.keyCode === 17) {
-                if (this.interactionData.onChangeToolsBlockerState && !this.thresholdWasModified) {
-                    this.interactionData.onChangeToolsBlockerState('keydown');
-                }
-                this.thresholdWasModified = false;
-            }
-        });
+        window.document.addEventListener('keyup', this.onKeyUp);
+        window.document.addEventListener('keydown', this.onKeyDown);
     }
 
     public transform(geometry: Geometry): void {
@@ -551,5 +541,10 @@ export class InteractionHandlerImpl implements InteractionHandler {
     public cancel(): void {
         this.release();
         this.onInteraction(null);
+    }
+
+    public destroy(): void {
+        window.document.removeEventListener('keyup', this.onKeyUp);
+        window.document.removeEventListener('keydown', this.onKeyDown);
     }
 }