+++ cvat/apps/engine/models.py
@@ -13,6 +13,7 @@ from django.conf import settings
 from django.contrib.auth.models import User
 from django.core.files.storage import FileSystemStorage
 from django.db import models
+from django.db.models.fields import FloatField
 from django.utils.translation import gettext_lazy as _
 from django.core.exceptions import SuspiciousFileOperation
 from rest_framework.exceptions import ValidationError
@@ -498,6 +499,7 @@ class Shape(models.Model):
     occluded = models.BooleanField(default=False)
     z_order = models.IntegerField(default=0)
     points = FloatArrayField()
+    rotation = FloatField(default=0)
 
     class Meta:
         abstract = True