+++ tests/cypress/integration/actions_tasks/case_71_shortcuts_window.js
@@ -17,7 +17,7 @@ context('Shortcuts window.', () => {
 
     describe(`Testing case "${caseId}"`, () => {
         it('Press "F1" from a task. Shortcuts window be visible. Closing the modal window by button "OK".', () => {
-            cy.get('body').trigger('keydown', { keyCode: keyCodeF1 });
+            cy.get('body').trigger('keydown', { keyCode: keyCodeF1, code: 'F1' });
             cy.get('.cvat-shortcuts-modal-window')
                 .should('exist')
                 .and('be.visible')
@@ -36,7 +36,7 @@ context('Shortcuts window.', () => {
 
         it('Open a job. Press "F1". Shortcuts window be visible. Closing the modal window by F1.', () => {
             cy.openJob();
-            cy.get('body').trigger('keydown', { keyCode: keyCodeF1 });
+            cy.get('body').trigger('keydown', { keyCode: keyCodeF1, code: 'F1' });
             cy.get('.cvat-shortcuts-modal-window')
                 .should('exist')
                 .and('be.visible')
@@ -49,7 +49,7 @@ context('Shortcuts window.', () => {
                             });
                     });
                 });
-            cy.get('body').trigger('keydown', { keyCode: keyCodeF1 });
+            cy.get('body').trigger('keydown', { keyCode: keyCodeF1, code: 'F1' });
             cy.get('.cvat-shortcuts-modal-window').should('not.be.visible');
         });
     });